add owner to tasks and groups, add authorization header to api requests

+++ api/src/routes/task.js
@@ -7,21 +7,21 @@ const router = express.Router()
 
 /* List of tasks */
 router.get('/tasks', verifyToken, (req, res) => {
-  TaskSchema.find()
+  TaskSchema.find({owner: req.userId})
     .then((tasks) => res.json(tasks))
     .catch((err) => res.status(400).json('Error: ' + err))
 })
 
 /* Get task by id */
 router.get('/task/:id', verifyToken, (req, res) => {
-  TaskSchema.findById({_id: req.params.id})
+  TaskSchema.findById({_id: req.params.id, owner: req.userId})
     .then((task) => res.json(task))
     .catch((err) => res.status(400).json('Error: ' + err))
 })
 
 /* Create new task */
 router.post('/task/add', verifyToken, (req, res) => {
-  const task = new TaskSchema(req.body)
+  const task = new TaskSchema({...req.body, owner: req.userId})
   task
     .save()
     .then((doc) => {
@@ -36,14 +36,17 @@ router.post('/task/add', verifyToken, (req, res) => {
 /* Edit task */
 router.post('/task/update/:id', verifyToken, (req, res) => {
   const {title, description} = req.body
-  TaskSchema.updateOne({_id: req.params.id}, {$set: {title, description}})
+  TaskSchema.updateOne(
+    {_id: req.params.id, owner: req.userId},
+    {$set: {title, description}}
+  )
     .then((data) => res.status(200).json(data))
     .catch((err) => res.status(400).json('Error: ' + err))
 })
 
 /* Delete task by id */
 router.delete('/task/remove/:id', verifyToken, (req, res) => {
-  TaskSchema.deleteOne({_id: req.params.id})
+  TaskSchema.deleteOne({_id: req.params.id, owner: req.userId})
     .then((data) => res.status(200).json(data))
     .catch((err) => res.status(400).json('Error: ' + err))
 })
